feat : #37 관리자 역할일 경우 수정 / 삭제 안되도록 기능 추가

+++ src/main/java/com/golapadeok/fluo/domain/role/service/RoleService.java
@@ -75,6 +75,11 @@ public class RoleService {
         Role role = this.roleRepository.findById(Long.valueOf(roleId))
                 .orElseThrow(() -> new RoleException(RoleErrorStatus.NOT_FOUND_ROLE));
 
+        // 역할의 이름이 관리자이고, 모든 권한을 갖고있으면(개수) 수정 못하도록 막기
+        if(role.getName().equals("관리자") && role.getRoleList().size() == 10) {
+            throw new RoleException(RoleErrorStatus.NOT_DELETE_ADMIN_ROLE);
+        }
+
         this.roleRepository.delete(role);
 
         String message = "역할이 삭제되었습니다.";
@@ -87,6 +92,11 @@ public class RoleService {
                 .orElseThrow(() -> new RoleException(RoleErrorStatus.NOT_FOUND_ROLE));
 
         log.info("role : {}", role.toString());
+
+        // 역할의 이름이 관리자이고, 모든 권한을 갖고있으면(개수) 수정 못하도록 막기
+        if(role.getName().equals("관리자") && role.getRoleList().size() == 10) {
+            throw new RoleException(RoleErrorStatus.NOT_UPDATE_ADMIN_ROLE);
+        }
 
         // 역할 업데이트
         role.updateRole(request);
